+++ share/dotfiles/.config/hypr/scripts/wallpaper-cache.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
-generated_versions="$HOME/.config/ml4w/cache/wallpaper-generated"
+
+ml4w_cache_folder="$HOME/.cache/ml4w/hyprland-dotfiles"
+
+generated_versions="$ml4w_cache_folder/wallpaper-generated"
+
 rm $generated_versions/*
 echo ":: Wallpaper cache cleared"
 notify-send "Wallpaper cache cleared"