+++ share/dotfiles/.config/ml4w/scripts/updates.sh
@@ -7,6 +7,17 @@
 #       |_|                              
 #  
 
+# Check if command exists
+_checkCommandExists() {
+    cmd="$1"
+    if ! command -v "$cmd" >/dev/null; then
+        echo 1
+        return
+    fi
+    echo 0
+    return
+}
+
 script_name=$(basename "$0")
 
 # Count the instances
@@ -24,42 +35,34 @@ fi
 threshhold_green=0
 threshhold_yellow=25
 threshhold_red=100
-install_platform="$(cat ~/.config/ml4w/settings/platform.sh)"
 
-# Check if platform is supported
-case $install_platform in
-    arch)
-        aur_helper="$(cat ~/.config/ml4w/settings/aur.sh)"
+# ----------------------------------------------------- 
+# Check for updates
+# ----------------------------------------------------- 
 
-        # ----------------------------------------------------- 
-        # Calculate available updates
-        # ----------------------------------------------------- 
+# Arch
+if [[ $(_checkCommandExists "pacman") == 0 ]]; then
+    aur_helper="$(cat ~/.config/ml4w/settings/aur.sh)"
 
-        # flatpak remote-ls --updates
+    check_lock_files() {
+        local pacman_lock="/var/lib/pacman/db.lck"
+        local checkup_lock="${TMPDIR:-/tmp}/checkup-db-${UID}/db.lck"
 
-        # -----------------------------------------------------------------------------
-        # Check for pacman or checkupdates-with-aur database lock and wait if necessary
-        # -----------------------------------------------------------------------------
-        check_lock_files() {
-            local pacman_lock="/var/lib/pacman/db.lck"
-            local checkup_lock="${TMPDIR:-/tmp}/checkup-db-${UID}/db.lck"
+        while [ -f "$pacman_lock" ] || [ -f "$checkup_lock" ]; do
+            sleep 1
+        done
+    }
 
-            while [ -f "$pacman_lock" ] || [ -f "$checkup_lock" ]; do
-                sleep 1
-            done
-        }
+    check_lock_files
 
-        check_lock_files
-
-        updates=$(checkupdates-with-aur | wc -l)
-    ;;
-    fedora)
-        updates=$(dnf check-update -q | grep -c ^[a-z0-9])
-    ;;
-    *)
-        updates=0
-    ;;
-esac
+    updates=$(checkupdates-with-aur | wc -l)
+# Fedora
+elif [[ $(_checkCommandExists "dnf") == 0 ]]; then
+    updates=$(dnf check-update -q | grep -c ^[a-z0-9])
+# Others
+else
+    updates=0
+fi
 
 # ----------------------------------------------------- 
 # Output in JSON format for Waybar Module custom-updates
@@ -75,8 +78,10 @@ if [ "$updates" -gt $threshhold_red ]; then
     css_class="red"
 fi
 
-if [ "$updates" -gt $threshhold_green ]; then
-    printf '{"text": "%s", "alt": "%s", "tooltip": "Click to update your system", "class": "%s"}' "$updates" "$updates" "$css_class"
-else
-    printf '{"text": "0", "alt": "0", "tooltip": "No updates available", "class": "green"}'
+if [ "$updates" != 0 ]; then
+    if [ "$updates" -gt $threshhold_green ]; then
+        printf '{"text": "%s", "alt": "%s", "tooltip": "Click to update your system", "class": "%s"}' "$updates" "$updates" "$css_class"
+    else
+        printf '{"text": "0", "alt": "0", "tooltip": "No updates available", "class": "green"}'
+    fi
 fi