+++ share/packages/fedora/profiles/addons.sh
@@ -1,7 +1,6 @@
 packages=(
   "firefox"
   "nautilus"
-  "nautilus-open-any-terminal"
   "gnome-text-editor"
   "gnome-calculator"
   "mpv"