+++ share/packages/fedora/profiles/default.sh
@@ -44,7 +44,6 @@ packages=(
     "grimblast-git"
     "bibata-cursor-theme"
     "fontawesome-6-free-fonts"
-    "mozilla-fira-sans-fonts"
     "fira-code-fonts"
     "flatpak"
     "oh-my-posh"